+++ tests/test_reader_read_header.py
@@ -43,3 +43,21 @@ def test_read_bgzip():
     assert str(r.header.lines[-1]) == EXPECTED
     assert r.header.samples
     assert r.header.samples.names == ['NA00001', 'NA00002', 'NA00003']
+
+
+def test_read_text_no_samples():
+    path = os.path.join(os.path.dirname(__file__), 'vcfs/from_vcf43_nosamples.vcf')
+    r = reader.Reader.from_path(path)
+    assert r.parser
+    assert r.header
+    assert len(r.header.lines) == 18
+    EXPECTED = "HeaderLine('fileformat', 'VCFv4.3')"
+    assert str(r.header.lines[0]) == EXPECTED
+    EXPECTED = (
+        "FormatHeaderLine('FORMAT', '<ID=HQ,Number=2,Type=Integer,"
+        "Description=\"Haplotype Quality\">', OrderedDict([('ID', 'HQ'), "
+        "('Number', 2), ('Type', 'Integer'), ('Description', "
+        "'Haplotype Quality')]))")
+    assert str(r.header.lines[-1]) == EXPECTED
+    assert r.header.samples
+    assert r.header.samples.names == []